backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sse-starlette==2.1.0
playwright==1.44.0
google-generativeai==0.7.2
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
aiofiles==23.2.1
python-multipart==0.0.9
